requirements.txt
langgraph
langchain-google-genai
langchain-core
yfinance
pandas
pandas_ta
python-dotenv
termcolor